Configration
Initialized the Rails API with the React App. We can now get all Users and display on the client.

We're using MySQL for the DB, and rspec for the api specs.

Also added some other dev dependencies such as factory-bot, database-cleaner and rubocop.

This is working nice now and the client can communicate with the Rails API.